refactor

+++ src/main/java/com/example/inkrhyme/repositories/IPoems.java
@@ -3,5 +3,5 @@ package com.example.inkrhyme.repositories;
 import com.example.inkrhyme.model.CPoems;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPoems extends JpaRepository <CPoems, Integer>{
+public interface IPoems extends JpaRepository <CPoems, Long>{
 }
